pyproject.toml
# Axiom-OS Configuration
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "axiom-os"
version = "0.1.0"
description = "Physics-Aware Neural Network Framework with Physical Anchoring"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Yuze Chen", email = "yuzechen83@example.com"}
]
keywords = ["physics-informed", "neural-network", "pin", "machine-learning", "scientific-computing"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Physics",
]
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "scikit-learn>=1.0.0",
    "matplotlib>=3.7.0",
    "typing-extensions>=4.5.0",
    "python-dotenv>=1.0.0",
    "requests>=2.28.0",
]

[project.optional-dependencies]
full = [
    "streamlit>=1.28.0",
    "openai>=1.0.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "astroquery>=0.4.6",
    "astropy>=5.0.0",
    "givernylocal>=3.2.0",
    "trimesh>=4.0.0",
    "duckduckgo-search>=4.0.0",
    "beautifulsoup4>=4.12.0",
    "trafilatura>=1.6.0",
]
mujoco = [
    "gymnasium>=0.29.0",
    "mujoco>=3.0.0",
    "imageio>=2.31.0",
    "stable-baselines3>=2.0.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "pre-commit>=3.5.0",
]
docs = [
    "sphinx>=7.2.0",
    "sphinx-rtd-theme>=2.0.0",
    "nbsphinx>=0.9.0",
]

[project.urls]
Homepage = "https://github.com/yuzechen83-crypto/axiom-os"
Documentation = "https://axiom-os.readthedocs.io"
Repository = "https://github.com/yuzechen83-crypto/axiom-os"
Issues = "https://github.com/yuzechen83-crypto/axiom-os/issues"

[project.scripts]
axiom = "axiom_os.cli.main:main"
axiom-main = "axiom_os.main:main"
axiom-ui = "axiom_os.agent.chat_ui:run_ui"
axiom-api = "axiom_os.api.server:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["axiom_os*"]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers"
testpaths = ["tests"]
pythonpath = ["."]
markers = [
    "slow: marks tests as slow",
    "gpu: marks tests that require GPU",
    "integration: marks tests as integration tests",
]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
